chore: separated image stack and spec table
